requirements.txt
flask
pandas
numpy
scikit-learn
scikit-surprise
seaborn
matplotlib